% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.log.growth.model.R
\name{bru_get_mapper.log_growth_model}
\alias{bru_get_mapper.log_growth_model}
\title{Mapper function for internal use}
\usage{
\method{bru_get_mapper}{log_growth_model}(model, ...)
}
\description{
Mapper function for internal use
}
