% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.timeonly.model.R
\name{bru_get_mapper.time_only_model}
\alias{bru_get_mapper.time_only_model}
\title{Mapper function for internal use}
\usage{
\method{bru_get_mapper}{time_only_model}(model, ...)
}
\description{
Mapper function for internal use
}
