% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.cgeneric.loggrow.model.R,
%   R/define.log.growth.model.R
\name{define.loggrow.model}
\alias{define.loggrow.model}
\alias{define.cgeneric.loggrow.model}
\title{Define logistic growth model in cgeneric}
\usage{
define.cgeneric.loggrow.model(
  linpoint,
  smesh,
  tmesh,
  step.size,
  prior.mean,
  prior.precision,
  priors = NULL,
  grad = NULL,
  initial.growth = NULL,
  initial.carry.cap = NULL,
  initial.move.const = NULL,
  initial.log.sigma = NULL,
  debug = NULL
)

define.loggrow.model(
  linpoint,
  smesh,
  tmesh,
  step.size,
  prior.mean,
  prior.precision,
  priors = NULL,
  grad = NULL,
  initial.growth = NULL,
  initial.carry.cap = NULL,
  initial.move.const = NULL,
  initial.log.sigma = NULL
)
}
\arguments{
\item{linpoint}{Linearisation point}

\item{smesh}{spatial mesh created with fm_mesh_2d_inla}

\item{tmesh}{mesh over equally spaced time pointes created with fm_mesh_1d}

\item{step.size}{difference between each of the time points}

\item{prior.mean}{estimated mean for first year of data. Helper function for calculating coming soon}

\item{prior.precision}{uncertainty for estimated mean of first year of data. Helper function for calculating coming soon}

\item{priors}{named list of prior parameters, named \code{cc} (carrying capacity), \code{growth}, \code{move}, \code{sigma}.
Each is a two element vector containing the mean and variance for each parameter.}

\item{initial.growth, initial.carry.cap, initial.move.const, initial.log.sigma}{Starting values for the growth,
\emph{log} carrying capacity, movement constant and \emph{log} standard deviation}
}
\value{
INLA cgeneric model packageCheck

INLA rgeneric model
}
\description{
Defines latent model and mapper for log.growth.regeneric

Defines latent model and mapper for log.growth.regeneric
}
