% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.timeonly.model.R
\name{define.loggrow.time.model}
\alias{define.loggrow.time.model}
\title{Define non spatial logistic growth model}
\usage{
define.loggrow.time.model(
  linpoint,
  tmesh,
  step.size,
  prior.mean,
  prior.precision,
  priors = NULL,
  initial.growth = NULL,
  initial.carry.cap = NULL,
  initial.log.sigma = NULL
)
}
\arguments{
\item{linpoint}{Linearisation point}

\item{tmesh}{mesh over equally spaced time pointes created with fm_mesh_1d}

\item{step.size}{difference between each of the time points}

\item{prior.mean}{estimated mean for first year of data. Helper function for calculating coming soon}

\item{prior.precision}{uncertainty for estimated mean of first year of data. Helper function for calculating coming soon}

\item{priors}{named list of prior parameters, named \code{cc} (carrying capacity), \code{growth}, \code{sigma}. 
Each is a two element vector containing the mean and variance for each parameter.}

\item{initial.growth, initial.carry.cap, initial.log.sigma}{Starting values for the \emph{log} growth, 
\emph{log} carrying capacity and \emph{log} standard deviation}
}
\value{
INLA rgeneric model
}
\description{
Defines latent model and mapper for log_growth_time
}
