% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper functions for interpreting.R
\name{get.initial.mean.var}
\alias{get.initial.mean.var}
\title{Extract prior mean and variance}
\usage{
get.initial.mean.var(fit, smesh)
}
\arguments{
\item{fit}{A fitted inlabru model of class "bru".}

\item{smesh}{Spatial mesh used to fit model}
}
\value{
named list of prior mean and variance
}
\description{
A helper function for getting prior mean and variance from an \strong{intercept and smoothing term only} inlabru model
fitted to first year of data. For more complicated models (e.g. including covariates or thinning),
the user needs to calculate these
}
