% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_of_linpoint.R
\name{gradient_of_linpoint}
\alias{gradient_of_linpoint}
\title{Numerically approximate gradient}
\usage{
gradient_of_linpoint(linpoint, smesh, tmesh)
}
\description{
Numerically approximate gradient
}
