% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.fit.R
\name{iterate.fit.custom}
\alias{iterate.fit.custom}
\title{Deprecated alias for \code{iterate.fit}.}
\usage{
iterate.fit.custom(
  formula,
  data,
  family,
  smesh,
  tmesh,
  samplers,
  prior.mean,
  prior.precision,
  max.iter = 100,
  gamma = 0.5,
  stop.crit = 0.05,
  priors = NULL,
  initial.linpoint = NULL,
  initial.growth = 0.5,
  initial.carry.cap = 1000,
  initial.move.const = 0.5,
  initial.sigma = log(1.5),
  method = "cgeneric",
  update.rule = 2,
  debug = F,
  options = NULL,
  saveall = T,
  weights = NULL,
  domain = NULL,
  early.stop = F,
  ...
)
}
\description{
Deprecated alias for \code{iterate.fit}.
}
