% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.fit.lgcp.R
\name{iterate.fit.lgcp}
\alias{iterate.fit.lgcp}
\title{Fit a logistic growth model with LGCP observations}
\usage{
iterate.fit.lgcp(
  data,
  smesh,
  tmesh,
  samplers,
  prior.mean,
  prior.variance,
  max.iter = 100,
  gamma = 0.5,
  stop.crit = 0.05,
  priors = NULL,
  initial.linpoint = NULL,
  initial.growth = 1,
  initial.carry.cap = 100,
  initial.move.const = 1,
  initial.log.sigma = log(1.5),
  verbose = F,
  saveall = T
)
}
\arguments{
\item{data}{spatial dataframe  observations}

\item{smesh}{spatial mesh created with \code{fm_mesh_2d_inla}}

\item{tmesh}{mesh over equally spaced time points created with \code{fm_mesh_1d}}

\item{samplers}{sampling area for supplying to inlabru}

\item{prior.mean}{estimated mean for first year of data. Helper function for calculating coming soon}

\item{prior.variance}{uncertainty for estimated mean of first year of data. Helper function for calculating coming soon}

\item{max.iter}{maximum iterations to attempt}

\item{gamma}{dampening parameter for update rule}

\item{stop.crit}{stopping criteria for linearisation point update rule. Stop updating if mean(abs(new_linearisation_point-old_linearisation_point))<=stop.crit}

\item{priors}{named list of prior parameters, named \code{cc} (carrying capacity), \code{growth}, \code{move}, \code{sigma}. 
Each is a two element vector containing the mean and variance for parameter.}

\item{initial.linpoint}{Optional. Starting guess for the linearisation point. If NULL, will be estimated within function}

\item{initial.growth, initial.carry.cap, initial.move.const, initial.log.sigma}{Starting values for the growth, 
\emph{log} carrying capacity, movement constant and \emph{log} standard deviation}

\item{verbose}{logical supplied to INLA}
}
\value{
list containing final model fit, number of iterations \code{n}, matrix of all past linearisation points and list of all past model fits.
}
\description{
Iteratively fits the spatial logistic growth model.
}
