% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.fit.R
\name{iterate.fit}
\alias{iterate.fit}
\title{Fit a logistic growth model with any observation type}
\usage{
iterate.fit(
  formula,
  data,
  family,
  smesh,
  tmesh,
  samplers,
  prior.mean,
  prior.precision,
  max.iter = 100,
  gamma = 0.5,
  stop.crit = 0.05,
  priors = NULL,
  initial.linpoint = NULL,
  initial.growth = 0.5,
  initial.carry.cap = 1000,
  initial.move.const = 0.5,
  initial.sigma = log(1.5),
  method = "cgeneric",
  update.rule = 2,
  debug = F,
  options = NULL,
  saveall = T,
  weights = NULL,
  domain = NULL,
  early.stop = F,
  ...
)
}
\arguments{
\item{formula}{an \code{inlabru} style specification of the latent components.
Should contain all elements of the linear predictor
except the spatial logistic growth term, which is added internally.
By default includes an intercept term, include -1 to remove the intercept. See also \code{\link{inlabru::bru_component()}}}

\item{data}{dataframe of the observations}

\item{family}{model family to use for observation process.
Common options are "cp" for LGCP, "poisson" and "nbinomial". Run \code{inla.list.models()} for all options}

\item{smesh}{spatial mesh created with fm_mesh_2d_inla}

\item{tmesh}{mesh over equally spaced time pointes created with fm_mesh_1d}

\item{samplers}{sampling area for supplying to inlabru}

\item{prior.mean}{estimated mean for first year of data. Helper function for calculating coming soon}

\item{prior.precision}{uncertainty for estimated mean of first year of data. Helper function for calculating coming soon}

\item{max.iter}{maximum iterations to attempt}

\item{gamma}{dampening parameter for update rule}

\item{stop.crit}{stopping criteria for linearisation point update rule. Stop updating if mean(abs(new_linearisation_point-old_linearisation_point))<=stop.crit}

\item{priors}{named list of prior parameters, named \code{cc} (carrying capacity), \code{growth}, \code{move}, \code{sigma}.
Each is a two element vector containing the mean on the log scale and standard devitation for the log normal prior}

\item{initial.linpoint}{Optional. Starting guess for the linearisation point. If NULL, will be estimated within function}

\item{initial.growth, initial.carry.cap, initial.move.const, initial.sigma}{Starting values for the growth,
carrying capacity, movement constant and standard deviation. Growth is on the real scale, the others on the log scale}

\item{verbose}{logical supplied to INLA}
}
\value{
list containing final model fit, number of iterations \code{n}, matrix of all past linearisation points and list of all past model fits.
}
\description{
Fits the spatial logistic growth model with any likelihood observation process
(of those supported by INLA) and a user defined formula.
}
