% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.fit.timeonly.R
\name{iterate.timeonly}
\alias{iterate.timeonly}
\title{Fit a non spatial logistic growth model with Poisson observations}
\usage{
iterate.timeonly(
  data,
  tmesh,
  step.size,
  prior.mean,
  prior.precision,
  max.iter = 100,
  gamma = 0.75,
  stop.crit = 0.05,
  priors = NULL,
  initial.linpoint = NULL,
  initial.growth = 1,
  initial.carry.cap = 0.05,
  initial.log.sigma = log(1.5),
  verbose = F,
  family,
  domain = NULL
)
}
\arguments{
\item{data}{spatial dataframe of gaussian observations, with column y = log(observed value)}

\item{tmesh}{mesh over equally spaced time points created with \code{fm_mesh_1d}}

\item{prior.mean}{estimated mean for first year of data.}

\item{prior.precision}{uncertainty for estimated mean of first year of data.}

\item{max.iter}{maximum iterations to attempt}

\item{gamma}{dampening parameter for update rule}

\item{stop.crit}{stopping criteria for linearisation point update rule. Stop updating if mean(abs(new_linearisation_point-old_linearisation_point))<=stop.crit}

\item{priors}{named list of prior parameters, named \code{cc} (carrying capacity), \code{growth}, \code{sigma}.
Each is a two element vector containing the mean and variance for parameter.}

\item{initial.linpoint}{Optional. Starting guess for the linearisation point. If NULL, will be estimated within function}

\item{initial.growth, initial.carry.cap, initial.log.sigma}{Starting values for the growth,
\emph{log} carrying capacity, movement constant and \emph{log} standard deviation}

\item{verbose}{logical supplied to INLA}
}
\value{
list containing final model fit, number of iterations \code{n}, matrix of all past linearisation points and list of all past model fits.
}
\description{
Iteratively fits the spatial logistic growth model.
}
