% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.growth.rgeneric.R
\name{log_growth_rgeneric}
\alias{log_growth_rgeneric}
\title{rgeneric Spatial Logistic Growth}
\usage{
log_growth_rgeneric(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL
)
}
\description{
The rgeneric function implementing the 2D spatial logistic growth function.
For use within INLA only. See inla.doc('rgeneric') for more
}
