% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.growth.time.R
\name{log_growth_time}
\alias{log_growth_time}
\title{rgeneric Time only Logistic Growth (not spatial)}
\usage{
log_growth_time(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL
)
}
\description{
The rgeneric function implementing the time only (non-spatial) logistic growth function. 
For use within INLA only. See inla.doc('rgeneric') for more
}
