% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_growth_helpers.R
\name{logit.growth}
\alias{logit.growth}
\title{Non spatial 1 year logistic growth function
Deterministic function for calculating population size after 1 time step of logistic growth}
\usage{
logit.growth(x, r, k)
}
\arguments{
\item{x}{numeric or vector starting population}

\item{r}{numeric or vector growth rate}

\item{k}{numeric or vector carrying capacity}
}
\value{
population after 1 time step of logistic growth
}
\description{
Non spatial 1 year logistic growth function
Deterministic function for calculating population size after 1 time step of logistic growth
}
