% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_growth_helpers.R
\name{logit.nest}
\alias{logit.nest}
\title{Nested logistic growth
Deterministic function for calculating population size over multiple time steps of growth}
\usage{
logit.nest(x0, r, k, n)
}
\arguments{
\item{x0}{numeric or vector starting population}

\item{r}{numeric or vector growth rate}

\item{k}{numeric or vector carrying capacity}

\item{n}{number of time steps to compute}
}
\value{
dataframe of population size and time
}
\description{
Nested logistic growth
Deterministic function for calculating population size over multiple time steps of growth
}
