% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cpp.R
\name{simulate_cpp}
\alias{simulate_cpp}
\title{Simulate Spatial Logistic Growth Data}
\usage{
simulate_cpp(
  growth,
  carry.cap,
  movement,
  sigma,
  initial.pop,
  initial.range,
  initial.sigma,
  timesteps,
  npoints = NULL,
  obs.sd = NULL,
  obs.prob = NULL,
  sample.type = "LGCP",
  ncores = 1,
  boundaries = c(0, 1),
  debug = F,
  max.edge = 0.05,
  nsurv = 3
)
}
\arguments{
\item{growth, k, movement, sigma}{parameters for spatial logistic growth model}

\item{timesteps}{Number of years to simulate}

\item{npoints}{For \code{sample.type="Gaussian"} only. Number of points to sample.}

\item{obs.sd}{For \code{sample.type="Gaussian"} only. Standard deviation of observation process}

\item{sample.type}{One of "LGCP" or "Gaussian". Default LGCP}

\item{ncores}{Optional. Number of cores to use if running in parallel.}

\item{boundaries}{Lower and upper limit of each side of the square. Defaults to sides of length 1}

\item{initial}{starting population size}
}
\value{
List containing underlying field (animal), observations (animal_obs), and
field defined on mesh nodes for debugging purposes (field).
}
\description{
Simulates observations and underlying field for spatial logistic growth in a square area.
}
