% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper functions for interpreting.R
\name{thetas2params}
\alias{thetas2params}
\title{Helper function for converting parameters from internal scale to more interpretable parameter scale}
\usage{
thetas2params(theta)
}
\arguments{
\item{theta}{vector of interal parameters}
}
\value{
named vector of user interpretable parameters
}
\description{
Helper function for converting parameters from internal scale to more interpretable parameter scale
}
