% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get precision matrix.R
\name{inla.logrowth.mean.precision}
\alias{inla.logrowth.mean.precision}
\title{Get precision matrix}
\usage{
inla.logrowth.mean.precision(smesh, tmesh, theta, initial)
}
\description{
Get precision matrix
}
