% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_growth_helpers.R
\name{logit.growth}
\alias{logit.growth}
\title{Non spatial 1 year logistic growth function
Deterministic function for calculating population size after 1 time step of logistic growth
@param x numeric or vector starting population
@param r numeric or vector growth rate
@param k numeric or vector carrying capacity
@returns population after 1 time step of logistic growth
@export}
\usage{
logit.growth(x, r, k)
}
\description{
Non spatial 1 year logistic growth function
Deterministic function for calculating population size after 1 time step of logistic growth
@param x numeric or vector starting population
@param r numeric or vector growth rate
@param k numeric or vector carrying capacity
@returns population after 1 time step of logistic growth
@export
}
