% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic growth helper functions.R
\name{logit.nest}
\alias{logit.nest}
\title{Nested logistic growth
Deterministic function for calculating population size over multiple time steps of growth
 @param x0 numeric or vector starting population
 @param r numeric or vector growth rate
 @param k numeric or vector carrying capacity
 @param n number of time steps to compute
 @returns dataframe of population size and time
 @export}
\usage{
logit.nest(x0, r, k, n)
}
\description{
Nested logistic growth
Deterministic function for calculating population size over multiple time steps of growth
 @param x0 numeric or vector starting population
 @param r numeric or vector growth rate
 @param k numeric or vector carrying capacity
 @param n number of time steps to compute
 @returns dataframe of population size and time
 @export
}
