% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get precision matrix.R
\name{simulate.loggrowth2}
\alias{simulate.loggrowth2}
\title{Functions for simulating growth}
\usage{
\method{simulate}{loggrowth2}(
  growth,
  k,
  movement,
  sigma,
  initial,
  timesteps,
  npoints = NULL,
  obs.sd = NULL,
  sample.type = "LGCP",
  ncores = 1,
  boundaries = c(0, 1)
)
}
\description{
Functions for simulating growth
}
